+++ docs/changelog.md
@@ -1,6 +1,7 @@
 ## v0.20.x
 
 - Adds a `GalaxyPlot` for plotting in galactic coordinates
+- Adds label support to lines with automatic angle adjustment
 
 
 ## v0.19.x